refactor: split engine-build
