+++ packages/engine-cli/src/index.ts
@@ -4,9 +4,8 @@ export {
     runEngine,
     loadEngineConfig,
     resolveRuntimeOptions,
-    runLocalNodeManager,
-    readMetadataFiles,
 } from './engine-build';
-
+export { runLocalNodeManager } from './run-local-mode-manager';
+export { readMetadataFiles } from './read-metadata-files';
 export { ManagedRunEngine } from './engine-app-manager';
 export type { IExecutableApplication, RunningFeature } from './types';
